requirements.txt
roboflow
python-dotenv
requests
supabase
pillow
fastapi